Phase 4: extract useSpecSubmissionFlow, the generation gate
The largest and most cohesive slice in the file, and the one that made
App.tsx hard to read: eleven pieces of state, a 90-line relocalization
effect, two async submit functions and five QA handlers, interleaved with
routing, catalog and theater code across 500 lines.

- useSpecSubmissionFlow.ts  submission status/error, the QA gate state
  (questions, pending spec, locale, builder, form key, parked answers),
  the language-switch re-ask effect, handleSubmitSpec, submitRefinedSpec,
  and handleQaComplete / handleQaCancel / handleQaAnswersChange /
  handleQaTitleChange / handleQaBuilderChange

The hook takes setSavedSpecs and setMyGamesRefreshKey because a landed
submission writes both — the saved-spec list drives the document title
and the refresh key drives the gallery, so neither belongs here. It calls
useTranslation itself: the refiner is told which language to write the
questions in, so i18n is an input to the flow, not a prop of it.

App.tsx no longer imports submissionApi, submissionErrors, pendingQa,
builderKind, gameTitle, conceptLength or visitTelemetry at all. With the
route-chunk boundary gone from between them, the two import blocks the
file used to carry are merged into one above the lazy() route chunks.

App.tsx 958 -> 624, from 1312 at the start of this series.
