+++ apps/web/src/App.tsx
@@ -1,28 +1,18 @@
-import { lazy, Suspense, useCallback, useEffect, useMemo, useRef, useState } from 'react';
+import { lazy, Suspense, useEffect, useMemo, useRef, useState } from 'react';
 import { useTranslation } from 'react-i18next';
 import { useCatalogData } from './useCatalogData.js';
 import { NavHeader } from './NavHeader.js';
 import { HeroPromptSection } from './HeroPromptSection.js';
 import { ArcadeCatalog } from './surfaces/catalog/ArcadeCatalog.js';
 import { UnpublishedPlayView } from './UnpublishedPlayView.js';
-import { CreatorQA, type QAQuestion } from './CreatorQA.js';
-import { deriveTitleFromConcept } from './gameTitle.js';
-import { MIN_CONCEPT_LENGTH } from './conceptLength.js';
+import { CreatorQA } from './CreatorQA.js';
 import { resolveCreateInitialPrompt } from './createInitialPrompt.js';
-import {
-  adminPath,
-  creatorPath,
-  navUpTarget,
-  playPath,
-  reviewPath,
-  studioPath,
-  studioWelcomePath,
-  studioConnectPath,
-} from './core/router.js';
+import { adminPath, creatorPath, navUpTarget, playPath, reviewPath, studioPath } from './core/router.js';
 import { RouteChunkBoundary } from './appRouteRecovery.js';
 import { useAppNavigation } from './useAppNavigation.js';
 import { useGameTheater } from './useGameTheater.js';
 import { StageOverlay } from './StageOverlay.js';
+import { useSpecSubmissionFlow } from './useSpecSubmissionFlow.js';
 import { StudioWelcomeView } from './surfaces/studio/StudioWelcomeView.js';
 import { StudioConnectWizard } from './surfaces/studio/StudioConnectWizard.js';
 import type { PublicCreatorProfile } from './creatorProfileApi.js';
@@ -43,6 +33,17 @@ import { SiteFooter } from './SiteFooter.js';
 import { resolveDocumentTitle } from './pageTitle.js';
 import { useDocumentTitle } from './useDocumentTitle.js';
 
+import { useActiveBuildCount } from './activeBuilds.js';
+import { getSavedSpecs, type SavedSpec } from './mySpecs.js';
+import { useAuth } from './AuthContext.js';
+import { AuthModal } from './AuthModal.js';
+import { BottomCta } from './BottomCta.js';
+import { ClosedBetaSplash } from './ClosedBetaSplash.js';
+import { BetaInvitePage } from './BetaInvitePage.js';
+import { AppLoadingScreen } from './AppLoadingScreen.js';
+import { ControllerView } from './surfaces/party/ControllerView.js';
+import { parseOAuthReturnParam } from './oauthReturn.js';
+
 // Deferred: an anonymous player playing a published game never has to pay for the
 // weight of the admin console, the studio (and everything it drags in — the code
 // editor, the remix panel), the review desk, or party's multiplayer client.
@@ -53,24 +54,8 @@ const CreatorStudioView = lazy(() =>
 const ReviewDesk = lazy(() => import('./surfaces/review/ReviewDesk.js').then((m) => ({ default: m.ReviewDesk })));
 const PartyPage = lazy(() => import('./surfaces/party/PartyPage.js').then((m) => ({ default: m.PartyPage })));
 
-import { submitSpec, refineSpec, type SubmissionApiError, type PlatformBuilderAvailability } from './submissionApi.js';
-import { submissionErrorKey } from './submissionErrors.js';
-import { useActiveBuildCount } from './activeBuilds.js';
-import { getSavedSpecs, saveSpec, type SavedSpec } from './mySpecs.js';
-import { saveLastBuilder, type BuilderKind } from './builderKind.js';
-import { clearPendingQa, loadPendingQa, savePendingQa, type PendingQaAnswers } from './pendingQa.js';
-import { useAuth } from './AuthContext.js';
-import { AuthModal } from './AuthModal.js';
-import { BottomCta } from './BottomCta.js';
-import { recordCreateStep, recordStudioStep } from './visitTelemetry.js';
-import { ClosedBetaSplash } from './ClosedBetaSplash.js';
-import { BetaInvitePage } from './BetaInvitePage.js';
-import { AppLoadingScreen } from './AppLoadingScreen.js';
-import { ControllerView } from './surfaces/party/ControllerView.js';
-import { parseOAuthReturnParam } from './oauthReturn.js';
-
 export function App() {
-  const { t, i18n } = useTranslation();
+  const { t } = useTranslation();
   const { user, loading: authLoading, privateBeta, publicPlaySlugs, showBetaWelcome, dismissBetaWelcome } = useAuth();
   const [isAuthModalOpen, setIsAuthModalOpen] = useState(false);
 
@@ -137,45 +122,23 @@ export function App() {
   // Paused while a game is on screen because the player covers the header.
   const activeBuildCount = useActiveBuildCount(myGamesRefreshKey, !stageContent);
 
-  // Greenfield submission state
-  // 'refining' is the spec-refiner call that precedes a submission — a few seconds
-  // during which nothing has been submitted yet, so the UI must not claim otherwise.
-  const [submissionStatus, setSubmissionStatus] = useState<'idle' | 'refining' | 'loading'>('idle');
-  const [submissionError, setSubmissionError] = useState<string | null>(null);
-  const submissionStatusRef = useRef(submissionStatus);
-  submissionStatusRef.current = submissionStatus;
-
-  // Clarifying-questions gate: a submission runs the spec refiner first, and when
-  // it returns questions the creator must answer them before generation proceeds.
-  // pendingSpec holds the spec awaiting those answers.
-  // Seeded from localStorage so a reload mid-round resumes instead of throwing the
-  // questions away and charging another refine to ask them again.
-  const restoredQa = useRef(loadPendingQa());
-  const [qaQuestions, setQaQuestions] = useState<QAQuestion[]>(restoredQa.current?.questions ?? []);
-  const [pendingSpec, setPendingSpec] = useState<{ title: string; concept: string; displayName: string } | null>(
-    restoredQa.current?.spec ?? null,
-  );
-  // Kept out of pendingSpec — too large for its localStorage-backed persistence.
-  const pendingReferenceImagesRef = useRef<string[] | undefined>(undefined);
-  // Language the parked questions were written in. Empty when an older blob never
-  // recorded one — that mismatch with the live UI language is what triggers a
-  // one-shot re-ask so English chips don't stick under a Polish chrome.
-  const [qaLocale, setQaLocale] = useState<string>(restoredQa.current?.locale ?? '');
-  // Who builds this round — parked with the confirm session so a reload keeps it.
-  const [qaBuilder, setQaBuilder] = useState<BuilderKind>(restoredQa.current?.builder ?? 'platform');
-  const qaBuilderRef = useRef(qaBuilder);
-  qaBuilderRef.current = qaBuilder;
-  // Whether the Gamedev.pl (platform) builder can be picked right now — reported by the
-  // hero's own quota poll, which already fetches this alongside the daily allowance.
-  const [platformBuilderAvailability, setPlatformBuilderAvailability] = useState<PlatformBuilderAvailability>();
-  // Bumped when questions are rewritten for a new language so CreatorQA remounts
-  // with empty answers — English chip labels must not survive as "selected" under
-  // Polish options that no longer match.
-  const [qaFormKey, setQaFormKey] = useState(0);
-  // Kept next to the QA state so the language-switch effect can clear it too.
-  const latestAnswersRef = useRef<PendingQaAnswers>(restoredQa.current?.answers ?? { selected: {}, custom: {} });
-
-  // Demo generator state
+  const {
+    submissionStatus,
+    submissionError,
+    pendingSpec,
+    qaQuestions,
+    qaBuilder,
+    qaFormKey,
+    latestAnswersRef,
+    platformBuilderAvailability,
+    setPlatformBuilderAvailability,
+    handleSubmitSpec,
+    handleQaComplete,
+    handleQaCancel,
+    handleQaAnswersChange,
+    handleQaTitleChange,
+    handleQaBuilderChange,
+  } = useSpecSubmissionFlow({ user, navigate, setIsAuthModalOpen, setSavedSpecs, setMyGamesRefreshKey });
 
   // Unpublished `/play/<slug>` title, reported once the document loads.
   const [unpublishedPlayTitle, setUnpublishedPlayTitle] = useState<string | null>(null);
@@ -246,105 +209,6 @@ export function App() {
     setGameTitle(null);
   }, [gameRouteSlug]);
 
-  // The static chrome follows the language switcher instantly; the AI questions do
-  // not — they were authored in whatever language the refine call used. Re-ask when
-  // the UI language changes mid-round (or a restored session was parked under a
-  // different language) so a Polish UI never keeps showing English chips.
-  const qaRelocalizingRef = useRef(false);
-  const pendingSpecRef = useRef(pendingSpec);
-  pendingSpecRef.current = pendingSpec;
-  useEffect(() => {
-    const parked = pendingSpecRef.current;
-    if (!parked) return;
-    const targetLocale = i18n.resolvedLanguage ?? i18n.language;
-    if (qaLocale === targetLocale) return;
-    // A real submit is in flight — don't yank the questions out from under it.
-    if (submissionStatusRef.current === 'loading') return;
-    if (qaRelocalizingRef.current) return;
-
-    let cancelled = false;
-    qaRelocalizingRef.current = true;
-    setSubmissionStatus('refining');
-    const concept = parked.concept;
-
-    async function relocalizeQa() {
-      try {
-        const refined = await refineSpec({ concept, locale: targetLocale });
-        if (cancelled) return;
-        // A real submit started while refine was in flight — drop the relocalization.
-        if (submissionStatusRef.current === 'loading') return;
-
-        const questions = refined.questions;
-        // Fail-open: when Vertex times out or errors, it returns empty questions.
-        // Keep the existing questions and parked session instead of wiping them into a name-only panel.
-        if (questions.length === 0) return;
-
-        // Prefer the live parked spec so a title edit mid-flight is not overwritten.
-        const liveSpec = pendingSpecRef.current;
-        if (!liveSpec) return;
-
-        // Preserve user-entered custom answers across questions matching by ID or index.
-        const oldCustom = latestAnswersRef.current.custom ?? {};
-        const oldQuestions = qaQuestions;
-        const preservedCustom: Record<string, string> = {};
-
-        questions.forEach((newQ, idx) => {
-          const customById = oldCustom[newQ.id];
-          const oldQ = oldQuestions[idx];
-          const customByIndex = oldQ ? oldCustom[oldQ.id] : undefined;
-          const val = customById || customByIndex;
-          if (val && val.trim()) {
-            preservedCustom[newQ.id] = val;
-          }
-        });
-
-        const newAnswers: PendingQaAnswers = { selected: {}, custom: preservedCustom };
-
-        setQaQuestions(questions);
-        setQaLocale(targetLocale);
-        latestAnswersRef.current = newAnswers;
-        // Drop restored answers so a remounted panel doesn't revive English selections.
-        if (restoredQa.current) {
-          restoredQa.current = {
-            ...restoredQa.current,
-            questions,
-            answers: newAnswers,
-            locale: targetLocale,
-            savedAt: Date.now(),
-          };
-        }
-        savePendingQa({
-          spec: liveSpec,
-          questions,
-          answers: newAnswers,
-          locale: targetLocale,
-          builder: qaBuilderRef.current,
-        });
-        setQaFormKey((key) => key + 1);
-      } catch {
-        // Keep the previous questions rather than blanking the panel on a blip.
-      } finally {
-        qaRelocalizingRef.current = false;
-        if (!cancelled && submissionStatusRef.current === 'refining') {
-          setSubmissionStatus('idle');
-        }
-      }
-    }
-
-    void relocalizeQa();
-    return () => {
-      cancelled = true;
-      // Strict Mode remounts (and a follow-up language flip) must be allowed to start
-      // a new call; leaving the guard latched would park the panel on "analyzing".
-      qaRelocalizingRef.current = false;
-    };
-    // Depend on the concept, not the whole pendingSpec object: title edits must not
-    // cancel and restart a language switch mid-flight. submissionStatus is read to
-    // skip during submit; listing it would cancel the relocalize when we flip to
-    // 'refining' ourselves.
-    // eslint-disable-next-line react-hooks/exhaustive-deps -- see above
-  }, [i18n.language, i18n.resolvedLanguage, pendingSpec?.concept, qaLocale]);
-
   // Menu navigation is scroll-to-section, but the sections only exist on the home
   // route — from a status page we have to go home first and scroll once the target
   // has mounted (the Games gallery may still be loading).
@@ -379,204 +243,6 @@ export function App() {
     }
     window.location.replace(oauthReturn);
   }, [authLoading, user]);
-
-  // The generation gate: before spending a submission we run the spec refiner. If it
-  // returns clarifying questions, generation pauses on the QA panel until they're
-  // answered; a clean spec submits straight through. A refiner error stops here too.
-  async function handleSubmitSpec(concept: string, displayName: string = '', referenceImages?: string[]) {
-    if (!user) {
-      // The wall between "wrote an idea" and "made an account". Everything before this
-      // is anonymous, so this is the only place that drop-off is visible at all.
-      recordCreateStep('signin_required');
-      setIsAuthModalOpen(true);
-      return;
-    }
-
-    const trimmedConcept = concept.trim();
-    if (!trimmedConcept) return;
-
-    // Catch a too-short concept before the fail-open refiner would.
-    if (trimmedConcept.length < MIN_CONCEPT_LENGTH) {
-      setSubmissionError(t('errors.conceptTooShort', { minLength: MIN_CONCEPT_LENGTH }));
-      return;
-    }
-
-    setSubmissionStatus('refining');
-    setSubmissionError(null);
-
-    let questions: QAQuestion[];
-    let suggestedTitle: string | undefined;
-    try {
-      const refined = await refineSpec({ concept: trimmedConcept, locale: i18n.language });
-      questions = refined.questions;
-      suggestedTitle = refined.suggestedTitle;
-    } catch {
-      // Fail-closed: stop here and let the creator retry, rather than falling
-      // through to a truncated title as if refinement had succeeded.
-      setSubmissionError(t('errors.refineFailed'));
-      setSubmissionStatus('idle');
-      return;
-    }
-
-    if (questions.length > 0) recordCreateStep('qa_shown');
-
-    // The confirm step always happens now, questions or not: it is where the game gets
-    // its name, and a build must not start without one the creator has seen.
-    const spec = {
-      title: suggestedTitle ?? deriveTitleFromConcept(trimmedConcept),
-      concept: trimmedConcept,
-      displayName: displayName.trim(),
-    };
-    const locale = i18n.resolvedLanguage ?? i18n.language;
-    pendingReferenceImagesRef.current = referenceImages;
-    setPendingSpec(spec);
-    setQaQuestions(questions);
-    setQaLocale(locale);
-    setQaBuilder('platform');
-    latestAnswersRef.current = { selected: {}, custom: {} };
-    savePendingQa({
-      spec,
-      questions,
-      answers: { selected: {}, custom: {} },
-      locale,
-      builder: 'platform',
-    });
-    setQaFormKey((key) => key + 1);
-    setSubmissionStatus('idle');
-  }
-
-  // Actually creates the submission (after the QA gate) and jumps to its status page.
-  async function submitRefinedSpec(title: string, concept: string, displayName: string, builder: BuilderKind) {
-    setSubmissionStatus('loading');
-    setSubmissionError(null);
-
-    try {
-      const response = await submitSpec({
-        title,
-        concept,
-        displayName: displayName || undefined,
-        // The agent is told this, so its progress updates arrive already written in
-        // the creator's language rather than machine-translated afterwards.
-        locale: i18n.language,
-        builder,
-        referenceImages: pendingReferenceImagesRef.current,
-      });
-
-      // Save to localStorage
-      const updatedSpecs = saveSpec({
-        token: response.token,
-        title,
-        concept,
-        createdAt: Date.now(),
-        ...(response.slug ? { slug: response.slug } : {}),
-      });
-      setSavedSpecs(updatedSpecs);
-      setMyGamesRefreshKey((key) => key + 1);
-      saveLastBuilder(response.token, builder);
-
-      setSubmissionStatus('idle');
-      recordCreateStep('submission_created', builder);
-      recordStudioStep('builder_chosen', builder);
-
-      // Only now is the QA panel done: it stayed up, in its submitting state, for the
-      // whole call. A no-op when the spec never went through the gate.
-      setQaQuestions([]);
-      setPendingSpec(null);
-      pendingReferenceImagesRef.current = undefined;
-      setQaLocale('');
-      setQaBuilder('platform');
-      clearPendingQa();
-
-      // Platform → welcome; self → connect chapter. Never auto-enter Studio.
-      const address = response.slug ?? response.token;
-      navigate(builder === 'platform' ? studioWelcomePath(address) : studioConnectPath(address));
-    } catch (err) {
-      const message = err instanceof Error ? err.message : t('errors.generic');
-      const apiErr = err instanceof Error ? (err as SubmissionApiError) : undefined;
-      setSubmissionError(
-        t(
-          submissionErrorKey({
-            message,
-            ...(apiErr?.status !== undefined ? { status: apiErr.status } : {}),
-            ...(apiErr?.category !== undefined ? { category: apiErr.category } : {}),
-          }),
-        ),
-      );
-      setSubmissionStatus('idle');
-    }
-  }
-
-  // The panel stays mounted until the submission actually lands. Clearing it first
-  // dropped the creator into blank space for however long the API took to create the
-  // issue — they had just clicked a button and the page answered by deleting itself.
-  // On failure it stays up with the error, so the answers survive a retry.
-  const handleQaComplete = async (finalConcept: string, title: string, builder: BuilderKind) => {
-    const spec = pendingSpec;
-    if (!spec) return;
-    // The name the creator settled on, which is the step that gates the build.
-    recordCreateStep('title_confirmed');
-    await submitRefinedSpec(title, finalConcept, spec.displayName, builder);
-  };
-
-  const handleQaCancel = () => {
-    setQaQuestions([]);
-    setPendingSpec(null);
-    pendingReferenceImagesRef.current = undefined;
-    setQaLocale('');
-    setQaBuilder('platform');
-    clearPendingQa();
-  };
-
-  // Every keystroke and chip lands in storage, so the round survives a reload at any
-  // point rather than only between questions.
-  const handleQaAnswersChange = useCallback(
-    (answers: PendingQaAnswers) => {
-      latestAnswersRef.current = answers;
-      if (!pendingSpec) return;
-      savePendingQa({
-        spec: pendingSpec,
-        questions: qaQuestions,
-        answers,
-        locale: qaLocale,
-        builder: qaBuilder,
-      });
-    },
-    [pendingSpec, qaQuestions, qaLocale, qaBuilder],
-  );
-
-  // The name is parked with the answers, for the same reason: an edited title is work,
-  // and a reload that kept the answers but silently restored the model's suggestion
-  // would be the one part of the panel that lies about having been saved.
-  const handleQaTitleChange = useCallback(
-    (title: string) => {
-      if (!pendingSpec) return;
-      const spec = { ...pendingSpec, title };
-      setPendingSpec(spec);
-      savePendingQa({
-        spec,
-        questions: qaQuestions,
-        answers: latestAnswersRef.current,
-        locale: qaLocale,
-        builder: qaBuilder,
-      });
-    },
-    [pendingSpec, qaQuestions, qaLocale, qaBuilder],
-  );
-
-  const handleQaBuilderChange = useCallback(
-    (builder: BuilderKind) => {
-      setQaBuilder(builder);
-      if (!pendingSpec) return;
-      savePendingQa({
-        spec: pendingSpec,
-        questions: qaQuestions,
-        answers: latestAnswersRef.current,
-        locale: qaLocale,
-        builder,
-      });
-    },
-    [pendingSpec, qaQuestions, qaLocale],
-  );
 
   // Unpublished `/play/<slug>` uses UnpublishedPlayView's own theater (not `stageContent`),
   // so hide Up the same way — Close / the error home link own escape there.
